requirements.txt
paramiko>=3.4.0
requests>=2.31.0
python-nmap>=0.7.1
bandit>=1.7.7
safety>=3.0.0
jinja2>=3.1.3
pyyaml>=6.0.1
cryptography>=42.0.0
